Raise error when bulk undeleting
Instead of doing nothing and treating the item as successful, we now raise a specific error.
This helps to determine if the bulk operation will trigger a history update when an item succeeds.
Otherwise, the UI will keep waiting for a history update that may never occur.
For the same reason, when purging an already purged item, we will touch the `update_time` but do nothing, so we notify the history to expect a *change*.

+++ lib/galaxy/webapps/galaxy/services/history_contents.py
@@ -1602,11 +1602,15 @@ class HistoryItemOperator:
 
     def _undelete(self, item: HistoryItemModel):
         if getattr(item, "purged", False):
-            return
+            raise exceptions.ItemDeletionException("This item has been permanently deleted and cannot be recovered.")
         manager = self._get_item_manager(item)
         manager.undelete(item, flush=self.flush)
 
     def _purge(self, item: HistoryItemModel, trans: ProvidesHistoryContext):
+        if getattr(item, "purged", False):
+            # TODO: remove this `update` when we can properly track the operation results to notify the history
+            item.update()
+            return
         if isinstance(item, HistoryDatasetCollectionAssociation):
             return self.dataset_collection_manager.delete(trans, "history", item.id, recursive=True, purge=True)
         self.hda_manager.purge(item, flush=self.flush)
